// libs/common/src/schemas/mongoose/admin/admin-permissions/admin-resources.enum.ts
export enum AdminResourcesEnum {
  ADMINS = 'admins',
  ADMIN_ROLES = 'adminRoles',
  USERS = 'users',
  PETS = 'pets',
  PET_BREEDS = 'petBreeds',
  PET_TYPES = 'petTypes',
  POSTS = 'posts',
  MODERATION_REPORTS = 'moderationReports',
  COMMENTS = 'comments',
  CITIES = 'cities',
  COUNTRIES = 'countries',
  AREAS = 'areas',
  APP_VERSIONS = 'appVersions',
  SYNC = 'sync',
  BRANCH_SERVICE_TYPE = 'branchServiceTypes',
  SERVICE_PROVIDERS = 'serviceProviders',
  APPOINTMENTS = 'appointments',
  EVENT_CATEGORIES = 'eventCategories',
  EVENT_FACILITIES = 'eventFacilities',
  MARKETING = 'marketing',
  LOST_FOUND_POSTS = 'lostFoundPosts',
  TOPICS = 'topics',
  BRANDS = 'brands',
  BRANCHES = 'branches',
  MEDICAL_SPECIALTY = 'medicalSpecialty',
  BRANCH_ACCESS_ROLE = 'branchAccessRole',
  PRODUCT_CATEGORIES = 'productCategories',
  PRODUCTS = 'products',
}